refactor(seat): SeatAssignmentService 트랜잭션 로직을 별도 빈으로 분리
SeatHoldService와 동일한 트랜잭션-락 경계 문제 해결.

기존 doAssignAndHold()는 같은 클래스 내부 호출(self-invocation)이라 @Transactional AOP 프록시가 적용되지 않는 문제도 함께 존재했음.

- SeatAssignmentTransactionalService를 별도 빈으로 분리하여 트랜잭션이 정상적으로 동작하고, 커밋 완료 후 락이 해제되도록 보장함.
- cleanupExistingHolds에서도 delete 후 flush()를 추가하여 JPA flush 순서 문제를 방지함.
