+++ Seat/src/main/java/com/goormgb/be/seat/recommendation/service/SeatAssignmentService.java
@@ -1,26 +1,14 @@
 package com.goormgb.be.seat.recommendation.service;
 
-import java.time.Clock;
-import java.time.Duration;
-import java.time.Instant;
-import java.util.List;
-
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.goormgb.be.global.exception.CustomException;
 import com.goormgb.be.global.exception.ErrorCode;
 import com.goormgb.be.seat.block.entity.Block;
 import com.goormgb.be.seat.block.repository.BlockRepository;
-import com.goormgb.be.seat.matchSeat.entity.MatchSeat;
-import com.goormgb.be.seat.matchSeat.repository.MatchSeatRepository;
-import com.goormgb.be.seat.recommendation.dto.internal.SemiGroup;
-import com.goormgb.be.seat.recommendation.dto.internal.SeatGroup;
 import com.goormgb.be.seat.recommendation.dto.response.SeatAssignmentResponse;
 import com.goormgb.be.seat.redis.SeatPreferenceRedisRepository;
 import com.goormgb.be.seat.redis.SeatSession;
-import com.goormgb.be.seat.seatHold.entity.SeatHold;
-import com.goormgb.be.seat.seatHold.repository.SeatHoldRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -33,27 +21,18 @@ import lombok.RequiredArgsConstructor;
  * <ol>
  *   <li>SeatSession에서 ticketCount 조회</li>
  *   <li>Redis 분산 락 획득 (block_lock:{blockId})</li>
- *   <li>진짜 연석 탐색</li>
- *   <li>(fallback) nearAdjacentToggle이 true이면 준연석 탐색</li>
- *   <li>좌석 상태를 BLOCKED로 변경</li>
- *   <li>SeatHold 생성 (5분 TTL)</li>
- *   <li>락 해제</li>
+ *   <li>트랜잭션 내에서 좌석 배정 + Hold 생성</li>
+ *   <li>트랜잭션 커밋 후 락 해제</li>
  * </ol>
  */
 @Service
 @RequiredArgsConstructor
 public class SeatAssignmentService {
 
-	private static final Duration HOLD_TTL = Duration.ofMinutes(5);
-
 	private final SeatPreferenceRedisRepository seatPreferenceRedisRepository;
 	private final BlockRepository blockRepository;
-	private final MatchSeatRepository matchSeatRepository;
-	private final SeatHoldRepository seatHoldRepository;
-	private final RealConsecutiveFinder realConsecutiveFinder;
-	private final SemiConsecutiveFinder semiConsecutiveFinder;
 	private final SeatBlockLock seatBlockLock;
-	private final Clock clock;
+	private final SeatAssignmentTransactionalService seatAssignmentTransactionalService;
 
 	public SeatAssignmentResponse assignAndHoldSeats(Long userId, Long matchId, Long blockId,
 		boolean nearAdjacentToggle) {
@@ -68,76 +47,10 @@ public class SeatAssignmentService {
 		}
 
 		try {
-			return doAssignAndHold(userId, matchId, blockId, block, requiredSeats, nearAdjacentToggle);
+			return seatAssignmentTransactionalService.assignAndHold(
+				userId, matchId, blockId, block, requiredSeats, nearAdjacentToggle);
 		} finally {
 			seatBlockLock.unlock(blockId);
 		}
-	}
-
-	@Transactional
-	protected SeatAssignmentResponse doAssignAndHold(
-		Long userId, Long matchId, Long blockId, Block block,
-		int requiredSeats, boolean nearAdjacentToggle
-	) {
-		// 기존 Hold 정리 (재요청 시)
-		cleanupExistingHolds(userId, matchId);
-
-		// 1. 진짜 연석 탐색
-		var realResult = realConsecutiveFinder.findBestRealConsecutive(matchId, blockId, requiredSeats);
-
-		if (realResult.isPresent()) {
-			SeatGroup seatGroup = realResult.get();
-			return holdSeats(userId, matchId, block, seatGroup.seats(), false);
-		}
-
-		// 2. 준연석 fallback (toggle이 켜져 있는 경우에만)
-		if (nearAdjacentToggle) {
-			var semiResult = semiConsecutiveFinder.findBestSemiConsecutive(matchId, blockId, requiredSeats);
-
-			if (semiResult.isPresent()) {
-				SemiGroup semiGroup = semiResult.get();
-				return holdSeats(userId, matchId, block, semiGroup.allSeats(), true);
-			}
-		}
-
-		throw new CustomException(ErrorCode.NO_CONSECUTIVE_SEAT_AVAILABLE);
-	}
-
-	private SeatAssignmentResponse holdSeats(
-		Long userId, Long matchId, Block block,
-		List<MatchSeat> seats, boolean semiConsecutive
-	) {
-		Instant expiresAt = clock.instant().plus(HOLD_TTL);
-
-		seats.forEach(MatchSeat::markBlocked);
-
-		List<SeatHold> holds = seats.stream()
-			.map(seat -> SeatHold.builder()
-				.matchSeatId(seat.getId())
-				.matchId(matchId)
-				.seatId(seat.getSeatId())
-				.userId(userId)
-				.expiresAt(expiresAt)
-				.build())
-			.toList();
-
-		seatHoldRepository.saveAll(holds);
-
-		return SeatAssignmentResponse.of(matchId, block, seats, expiresAt, semiConsecutive);
-	}
-
-	private void cleanupExistingHolds(Long userId, Long matchId) {
-		List<SeatHold> existingHolds = seatHoldRepository.findAllByUserIdAndMatchId(userId, matchId);
-
-		if (existingHolds.isEmpty()) {
-			return;
-		}
-
-		List<Long> matchSeatIds = existingHolds.stream().map(SeatHold::getMatchSeatId).toList();
-
-		List<MatchSeat> seatsToRelease = matchSeatRepository.findAllById(matchSeatIds);
-		seatsToRelease.forEach(MatchSeat::markAvailable);
-
-		seatHoldRepository.deleteAllByMatchSeatIdIn(matchSeatIds);
 	}
 }
